[fix] #163 - /api/v1/userRank/school 수정 작업 진행
- 응답 관련 수정 작업 진행
- 경쟁관련 race 관련 로직 삭제 진행! -> 기획에서 사라진 기능
