+++ src/main/java/gitbal/backend/api/userPage/controller/UserRankContorller.java
@@ -3,9 +3,8 @@ package gitbal.backend.api.userPage.controller;
 
 
 import gitbal.backend.api.userPage.dto.RegionRankRaceResponseDto;
-import gitbal.backend.api.userPage.dto.SchoolRankRaceResponseDto;
+import gitbal.backend.api.userPage.dto.SchoolRankResponseDto;
 import gitbal.backend.api.userPage.dto.UserRankMajorLanguageResponseDto;
-import gitbal.backend.api.userPage.dto.UserRankScoreResponseDto;
 import gitbal.backend.api.userPage.dto.UserRankingResponseDto;
 import gitbal.backend.api.userPage.service.UserRankService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -50,12 +49,12 @@ public class UserRankContorller {
 
     //TODO : 수정작업 진행 자신의 학교 몇등이고 어느 학교인지만 나오게
     @GetMapping("/school")
-    @Operation(summary = "유저랭크 학교 탭 부분(개발 완료)", description = "유저 랭크 페이지의 학교 탭의 정보를 위한 api")
+    @Operation(summary = "유저랭크 학교 탭 부분(8.11 수정-개발 완료)", description = "유저 랭크 페이지의 학교 탭의 정보를 위한 api")
     @ApiResponses(value = {
         @ApiResponse(responseCode = "200", description = "유저 학교 상황 정보 요청을 성공했습니다."),
         @ApiResponse(responseCode = "5xx", description = "유저 학교 상황 정보 요청을 실패했습니다.")
     })
-    public ResponseEntity<SchoolRankRaceResponseDto> schoolTab(@RequestParam String username){
+    public ResponseEntity<SchoolRankResponseDto> schoolTab(@RequestParam String username){
         return ResponseEntity.ok(userRankService.makeUserRankSchoolStatusByUsername(username));
     }
